requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic==0.34.0
openai==1.51.0
aiomysql==0.2.0
python-dotenv==1.0.1
chromadb==0.5.5
sentence-transformers==3.1.1
pypdf==4.3.1
pydantic==2.9.0
